8250_mid: arm Rx DMA on all ports with DMA continously
Instead of initiating Rx DMA when the first char arrives in the UART, arm Rx DMA
continously. DMA transfers automatically empty the Rx FIFO, preventing overruns.
As before, transfer terminates when the DMA buffer is full or when a 4 char
interchar gap is received (timeout). After timeout we arm immediately the DMA again.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>