serial: 8250_port: when using DMA do not split writes in TTY
By default tty_io write() splits large writes into 2K chunks. This
results to a single DMA transfer to split into 2 seperate transfers
which are not combined on the scatter gather list. As a result, under
heavy load or PM conditions an interchar gap occurs which can lead
to errors in the receiver.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>